ci: build packges before linting files
The react package depends on the core packge and needs
its source files and type declarations.

We need to build the packages before linting, otherwise
the react package cannot read the files of the core package.

Otherwise, the CI job will fail, because the bundled source
files of the core package do not exist.